Give delayed calls one due order, one command parser and the reference verb chain
Delayed TIMERF work diverged from CObjBase::r_Verb in five ways at once, and the
transport under it was narrower than upstream's.

Due order was per-object: the tick drained each object's ready jobs before moving
on, so a job due 10 seconds ago ran after one due 100ms ago purely because its
owner came later in the active set. Upstream keeps every timed object in one
time-sorted vector and ticks the selection in that order (CWorldTicker.cpp:1051).
Ready jobs are now gathered across all objects and sorted by due time, with the
collection order as the tie-break; each is still removed immediately before it
runs, so a save taken mid-callback still sees the ones that have not.

The payload split on whitespace alone, so "f_capture=37" queued a function named
"f_capture=37" that resolved to nothing. Upstream keeps the raw payload and runs
ParseKey over it when the timer fires (CTimedFunction.cpp:88, CScript.cpp:336),
whose Str_Parse separators are "=, \t" (CExpression.cpp:144). ScriptCommandLine
implements that split - separators plus the quote and bracket rules, only the
first one structural - and both ScheduleTimerF and the classic loader use it.

The dispatch itself stopped after the verb table. A single bool conflated "no
such verb" with "that verb refused", so a pack's [FUNCTION UNEQUIP] could pick up
a line the engine had just declined; upstream reaches the function branch only
when the table lookup misses (CObjBase.cpp:2134). TryExecuteCommand now reports
whether the NAME belongs to the verb table, cleared at the single fall-through
every switch ends in, so no per-case bookkeeping is needed. An unowned name then
reaches the script function and, failing that, a property assignment through the
default r_LoadVal branch (CScriptObj.cpp:1481) - "NAME after" used to do nothing.
The TOPOBJ./CONT./LINK. chain ran the same truncated order on its resolved
target; it now hands the remainder to that object's full verb line, so
"TOPOBJ.f_mark 37" no longer resolves the reference and then swallows the call.

Verbs needing the caller read the character off the client interface alone, so
UNEQUIP, SUMMONTO and CONTROL refused whenever no client was attached. Upstream
runs a delayed call with the top-level CHARACTER as SRC and reaches it through
pSrc->GetChar() (CTimedFunction.cpp:43, CItem.cpp:3574). ITextConsole gained that
accessor and the stand-in console for a clientless character answers it.

Numeric arguments saturated at the 32-bit bounds and the hex scan advanced over
ASCII digits only, so 2147483648 arrived as 2147483647 and "0A,2,3" delivered
0,0,0 - the stranded cursor took ARGN2 and ARGN3 down with it. ARGN1/2/3 are
int64 upstream (CScriptTriggerArgs.h:21); the transport is 64-bit end to end and
the engine's own int fields narrow at their own boundary through
ScriptNumber.ToEngineInt, saturating rather than wrapping. The number reader is
the GetSingle port: a leading '0' is the hex marker, then [0-9 A-F a-f], with the
significant-nibble width deciding the signed reinterpretation
(CExpression.cpp:666).

Finally, the 64-jobs-per-object cap dropped the 65th silently while the classic
loader counted it as restored. Upstream has no per-object limit
(CTimedFunctionHandler.cpp:103), so AddTimerF reports acceptance, a save bypasses
the cap because a save is authoritative, and a live refusal is logged.

Verified: 3095 tests pass (31 new in DelayedCallParity12XYTests).

+++ src/SphereNet.Core/Interfaces/ITriggerArgs.cs
@@ -8,8 +8,12 @@ public interface ITriggerArgs
     IScriptObj? Source { get; }
     IScriptObj? Object1 { get; }
     IScriptObj? Object2 { get; }
-    int Number1 { get; set; }
-    int Number2 { get; set; }
-    int Number3 { get; set; }
+    // Source-X CScriptTriggerArgs keeps ARGN1/2/3 as int64 (CScriptTriggerArgs.h:21).
+    // These were int, and InitFromRaw saturated a larger script value at
+    // int.MinValue/MaxValue, so a big counter or timestamp reached the script as a
+    // different number.
+    long Number1 { get; set; }
+    long Number2 { get; set; }
+    long Number3 { get; set; }
     string ArgString { get; set; }
 }
